tests: move CLITestCase to common folder
