+++ tests/Integrations/CLI/CakePHP/V2_8/CommonScenariosTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Integrations\CLI\CakePHP\V2_8;
 
 use DDTrace\Tests\Common\SpanAssertion;
-use DDTrace\Tests\Integrations\CLI\CLITestCase;
+use DDTrace\Tests\Common\CLITestCase;
 
 class CommonScenariosTest extends CLITestCase
 {
