+++ tests/Integrations/CLI/Custom/Autoloaded/CommonScenariosTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Integrations\CLI\Custom\Autoloaded;
 
 use DDTrace\Tests\Common\SpanAssertion;
-use DDTrace\Tests\Integrations\CLI\CLITestCase;
+use DDTrace\Tests\Common\CLITestCase;
 
 final class CommonScenariosTest extends CLITestCase
 {
